requirements.txt
blinker==1.6.3
click==8.1.7
flask==2.2.5
flask-sqlalchemy==3.0.5
greenlet==3.0.1
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
SQLAlchemy==2.0.22
typing-extensions==4.7.1
werkzeug==2.2.3
zipp==3.15.0
gunicorn==21.2.0
python-dotenv==0.21.1
psycopg2-binary==2.9.9
Flask-Migrate==4.0.5